chore: rename parser to argumentParser.ts
